+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # v4.10.0
 * Added: System property "metricRegistryStartConsoleReporter" to start the metric registry console reporter automatically
+* Added: Add fetch hint to exclude edge vertex ids or edge ids from edge references
 * Changed: Elasticsearch: Throw better exceptions when elements are missing in missing document helper
 * Fixed: Accumulo: don't queue events when no event listeners are attached